+++ mbed-client-pal/Source/Port/Reference-Impl/OS_Specific/FreeRTOS/Storage/FileSystem/pal_plat_fileSystem.c
@@ -289,11 +289,11 @@ palStatus_t pal_plat_fsFwrite(palFileDescriptor_t *fd, const void *buffer, size_
 }
 
 
-palStatus_t pal_plat_fsFseek(palFileDescriptor_t *fd, off_t offset, pal_fsOffset_t whence)
+palStatus_t pal_plat_fsFseek(palFileDescriptor_t *fd, int32_t offset, pal_fsOffset_t whence)
 {
     palStatus_t ret = PAL_SUCCESS;
     FRESULT status = FR_OK;
-    size_t  fatFSOffset = 0;
+    uint32_t  fatFSOffset = 0;
 
     if (CHK_FD_VALIDITY(*fd))
     {//Bad File Descriptor
@@ -354,7 +354,7 @@ palStatus_t pal_plat_fsFseek(palFileDescriptor_t *fd, off_t offset, pal_fsOffset
 }
 
 
-palStatus_t pal_plat_fsFtell(palFileDescriptor_t *fd, off_t * pos)
+palStatus_t pal_plat_fsFtell(palFileDescriptor_t *fd, int32_t * pos)
 {
     palStatus_t ret = PAL_SUCCESS;
